requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
flask>=3.0.0
requests>=2.31.0
tqdm>=4.65.0
faiss-cpu>=1.7.4
regex>=2023.0.0

